Move PopulateWithDefaultHeaderValues to MainForm.Movie, make private
Author still needs to be set locally
