+++ src/BizHawk.Client.Common/movie/MovieConversionExtensions.cs
@@ -4,17 +4,6 @@ using System.Linq;
 
 using BizHawk.Common;
 using BizHawk.Common.PathExtensions;
-using BizHawk.Emulation.Common;
-using BizHawk.Emulation.Cores.Arcades.MAME;
-using BizHawk.Emulation.Cores.Atari.Jaguar;
-using BizHawk.Emulation.Cores.Consoles.Nintendo.Ares64;
-using BizHawk.Emulation.Cores.Consoles.Nintendo.Gameboy;
-using BizHawk.Emulation.Cores.Consoles.Nintendo.NDS;
-using BizHawk.Emulation.Cores.Consoles.Sega.gpgx;
-using BizHawk.Emulation.Cores.Consoles.Sega.PicoDrive;
-using BizHawk.Emulation.Cores.Nintendo.NES;
-using BizHawk.Emulation.Cores.Nintendo.SubNESHawk;
-using BizHawk.Emulation.Cores.Sega.MasterSystem;
 
 namespace BizHawk.Client.Common
 {
@@ -181,127 +170,6 @@ namespace BizHawk.Client.Common
 
 			tas.Save();
 			return tas;
-		}
-
-		// TODO: This doesn't really belong here, but not sure where to put it
-		public static void PopulateWithDefaultHeaderValues(
-			this IMovie movie,
-			IEmulator emulator,
-			ISettingsAdapter settable,
-			IGameInfo game,
-			FirmwareManager firmwareManager,
-			string author)
-		{
-			movie.Author = author;
-			movie.EmulatorVersion = VersionInfo.GetEmuVersion();
-			movie.OriginalEmulatorVersion = VersionInfo.GetEmuVersion();
-			movie.SystemID = emulator.SystemId;
-
-			if (settable.HasSyncSettings)
-			{
-				movie.SyncSettingsJson = ConfigService.SaveWithType(settable.GetSyncSettings());
-			}
-
-			movie.GameName = game.FilesystemSafeName();
-			movie.Hash = game.Hash;
-			if (game.FirmwareHash != null)
-			{
-				movie.FirmwareHash = game.FirmwareHash;
-			}
-
-			if (emulator.HasBoardInfo())
-			{
-				movie.BoardName = emulator.AsBoardInfo().BoardName;
-			}
-
-			if (emulator.HasRegions())
-			{
-				var region = emulator.AsRegionable().Region;
-				if (region == DisplayType.PAL)
-				{
-					movie.HeaderEntries.Add(HeaderKeys.Pal, "1");
-				}
-			}
-
-			if (firmwareManager.RecentlyServed.Count != 0)
-			{
-				foreach (var firmware in firmwareManager.RecentlyServed)
-				{
-					var key = firmware.ID.MovieHeaderKey;
-					if (!movie.HeaderEntries.ContainsKey(key))
-					{
-						movie.HeaderEntries.Add(key, firmware.Hash);
-					}
-				}
-			}
-
-			if (emulator is NDS nds && nds.IsDSi)
-			{
-				movie.HeaderEntries.Add("IsDSi", "1");
-
-				if (nds.IsDSiWare)
-				{
-					movie.HeaderEntries.Add("IsDSiWare", "1");
-				}
-			}
-
-			if ((emulator is NES nes && nes.IsVS)
-				|| (emulator is SubNESHawk subnes && subnes.IsVs))
-			{
-				movie.HeaderEntries.Add("IsVS", "1");
-			}
-
-			if (emulator is IGameboyCommon gb)
-			{
-				//TODO doesn't IsCGBDMGMode imply IsCGBMode?
-				if (gb.IsCGBMode) movie.HeaderEntries.Add(gb.IsCGBDMGMode ? "IsCGBDMGMode" : "IsCGBMode", "1");
-			}
-
-			if (emulator is SMS sms)
-			{
-				if (sms.IsSG1000)
-				{
-					movie.HeaderEntries.Add("IsSGMode", "1");
-				}
-
-				if (sms.IsGameGear)
-				{
-					movie.HeaderEntries.Add("IsGGMode", "1");
-				}
-			}
-
-			if (emulator is GPGX gpgx && gpgx.IsMegaCD)
-			{
-				movie.HeaderEntries.Add("IsSegaCDMode", "1");
-			}
-
-			if (emulator is PicoDrive pico && pico.Is32XActive)
-			{
-				movie.HeaderEntries.Add("Is32X", "1");
-			}
-
-			if (emulator is VirtualJaguar jag && jag.IsJaguarCD)
-			{
-				movie.HeaderEntries.Add("IsJaguarCD", "1");
-			}
-
-			if (emulator is Ares64 ares && ares.IsDD)
-			{
-				movie.HeaderEntries.Add("IsDD", "1");
-			}
-
-			if (emulator is MAME mame)
-			{
-				movie.HeaderEntries.Add(HeaderKeys.VsyncAttoseconds, mame.VsyncAttoseconds.ToString());
-			}
-
-			if (emulator.HasCycleTiming())
-			{
-				movie.HeaderEntries.Add(HeaderKeys.CycleCount, "0");
-				movie.HeaderEntries.Add(HeaderKeys.ClockRate, "0");
-			}
-
-			movie.Core = emulator.Attributes().CoreName;
 		}
 
 		internal static string ConvertFileNameToTasMovie(string oldFileName)
